pyproject.toml
[project]
name = "data-transformation"
version = "0.1.0"
description = "Golden Record POC - End-to-End Data Deduplication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python==2.0.2",
    "pandas==2.1.3",
    "rapidfuzz==3.5.2",
    "python-dotenv==1.0.0",
    "tabulate==0.9.0",
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "main",
    "db_utils",
    "kafka_producer",
    "kafka_consumer",
    "silver_transformer",
    "dedup_engine",
    "decision_engine",
    "golden_merger",
    "query_db"
]

[tool.setuptools.packages.find]
include = ["golden_record_platform*"]
exclude = ["__pycache__*"]
